services and unit tests for services

+++ neu-fall22-pdp-hw4/src/portfolio/services/datastore/FileIOService.java
@@ -2,6 +2,7 @@ package portfolio.services.datastore;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.FileReader;
 import java.io.IOException;
@@ -13,10 +14,11 @@ import java.nio.charset.StandardCharsets;
  *
  */
 public class FileIOService implements IOService {
+
   @Override
-  public String read(String fileName) {
+  public String read(String fileName) throws IOException {
     String s;
-    try(BufferedReader br = new BufferedReader(new FileReader(fileName))) {
+    try (BufferedReader br = new BufferedReader(new FileReader(fileName))) {
       StringBuilder sb = new StringBuilder();
       String line = br.readLine();
 
@@ -26,14 +28,17 @@ public class FileIOService implements IOService {
         line = br.readLine();
       }
       s = sb.toString();
-    } catch (IOException e) {
-      throw new RuntimeException(e);
     }
     return s;
   }
 
   @Override
-  public boolean saveTo(String text, String fileName) {
+  public boolean saveTo(String text, String fileName) throws IllegalArgumentException {
+    File f = new File(fileName);
+    if (f.exists() && !f.isDirectory()) {
+      throw new IllegalArgumentException(
+          "There is a file or a directory exists with filename: " + fileName);
+    }
     try (Writer writer = new BufferedWriter(new OutputStreamWriter(
         new FileOutputStream(fileName), StandardCharsets.UTF_8))) {
       writer.write(text);
